requirements.txt
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
pyyaml==6.0.2
joblib==1.4.2
librosa==0.10.2.post1
soundfile==0.12.1
scipy==1.14.1
